+++ crates/edit_prediction/src/edit_prediction_tests.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::cursor_excerpt::editable_and_context_ranges_for_cursor_position;
 use crate::{compute_diff_between_snapshots, udiff::apply_diff_to_string, zeta1::MAX_EVENT_TOKENS};
 use client::{UserStore, test::FakeServer};
 use clock::{FakeSystemClock, ReplicaId};
@@ -19,7 +18,7 @@ use gpui::{
     http_client::{FakeHttpClient, Response},
 };
 use indoc::indoc;
-use language::{Buffer, Point, rust_lang};
+use language::{Buffer, Point};
 use lsp::LanguageServerId;
 use parking_lot::Mutex;
 use pretty_assertions::{assert_eq, assert_matches};
@@ -27,11 +26,7 @@ use project::{FakeFs, Project};
 use serde_json::json;
 use settings::SettingsStore;
 use std::{path::Path, sync::Arc, time::Duration};
-use util::{
-    path,
-    rel_path::rel_path,
-    test::{TextRangeMarker, marked_text_ranges_by},
-};
+use util::{path, rel_path::rel_path};
 use uuid::Uuid;
 use zeta_prompt::ZetaPromptInput;
 
@@ -2379,276 +2374,4 @@ fn test_compute_diff_between_snapshots(cx: &mut TestAppContext) {
 #[ctor::ctor]
 fn init_logger() {
     zlog::init_test();
-}
-
-struct TestCase {
-    name: &'static str,
-    marked_text: &'static str,
-    editable_token_limit: usize,
-    context_token_limit: usize,
-}
-
-#[gpui::test]
-fn test_editable_and_context_ranges(cx: &mut App) {
-    // Markers:
-    // ˇ = cursor position
-    // « » = expected editable range
-    // [ ] = expected context range
-    let test_cases = vec![
-        TestCase {
-            name: "cursor near end of function - expands to syntax boundaries",
-            marked_text: indoc! {r#"
-                [fn first() {
-                    let a = 1;
-                    let b = 2;
-                }
-
-                fn foo() {
-                «    let x = 1;
-                    let y = 2;
-                    println!("{}", x + y);ˇ
-                }»]
-            "#},
-            // 18 tokens - expands symmetrically then to syntax boundaries
-            editable_token_limit: 18,
-            context_token_limit: 35,
-        },
-        TestCase {
-            name: "cursor at function start - expands to syntax boundaries",
-            marked_text: indoc! {r#"
-                [fn before() {
-                «    let a = 1;
-                }
-
-                fn foo() {ˇ
-                    let x = 1;
-                    let y = 2;
-                    let z = 3;
-                }
-                »
-                fn after() {
-                    let b = 2;
-                }]
-            "#},
-            // 25 tokens - expands symmetrically then to syntax boundaries
-            editable_token_limit: 25,
-            context_token_limit: 50,
-        },
-        TestCase {
-            name: "tiny budget - just lines around cursor",
-            marked_text: indoc! {r#"
-                fn outer() {
-                [    let line1 = 1;
-                    let line2 = 2;
-                «    let line3 = 3;
-                    let line4 = 4;ˇ»
-                    let line5 = 5;
-                    let line6 = 6;]
-                    let line7 = 7;
-                }
-            "#},
-            // 12 tokens (~36 bytes) = just the cursor line with tiny budget
-            editable_token_limit: 12,
-            context_token_limit: 24,
-        },
-        TestCase {
-            name: "small function fits entirely",
-            marked_text: indoc! {r#"
-                [«fn foo() {
-                    let x = 1;ˇ
-                    let y = 2;
-                }»]
-            "#},
-            // Plenty of budget for this small function
-            editable_token_limit: 30,
-            context_token_limit: 60,
-        },
-        TestCase {
-            name: "context extends beyond editable",
-            marked_text: indoc! {r#"
-                [fn first() { let a = 1; }
-                «fn second() { let b = 2; }
-                fn third() { let c = 3; }ˇ
-                fn fourth() { let d = 4; }»
-                fn fifth() { let e = 5; }]
-            "#},
-            // Small editable, larger context
-            editable_token_limit: 25,
-            context_token_limit: 45,
-        },
-        // Tests for syntax-aware editable and context expansion
-        TestCase {
-            name: "cursor in first if-statement - expands to syntax boundaries",
-            marked_text: indoc! {r#"
-                [«fn before() { }
-
-                fn process() {
-                    if condition1 {
-                        let a = 1;ˇ
-                        let b = 2;
-                    }
-                    if condition2 {»
-                        let c = 3;
-                        let d = 4;
-                    }
-                    if condition3 {
-                        let e = 5;
-                        let f = 6;
-                    }
-                }
-
-                fn after() { }]
-            "#},
-            // 35 tokens allows expansion to include function header and first two if blocks
-            editable_token_limit: 35,
-            // 60 tokens allows context to include the whole file
-            context_token_limit: 60,
-        },
-        TestCase {
-            name: "cursor in middle if-statement - expands to syntax boundaries",
-            marked_text: indoc! {r#"
-                [fn before() { }
-
-                fn process() {
-                    if condition1 {
-                        let a = 1;
-                «        let b = 2;
-                    }
-                    if condition2 {
-                        let c = 3;ˇ
-                        let d = 4;
-                    }
-                    if condition3 {
-                        let e = 5;»
-                        let f = 6;
-                    }
-                }
-
-                fn after() { }]
-            "#},
-            // 40 tokens allows expansion to surrounding if blocks
-            editable_token_limit: 40,
-            // 60 tokens allows context to include the whole file
-            context_token_limit: 60,
-        },
-        TestCase {
-            name: "cursor near bottom of long function - editable expands toward syntax, context reaches function",
-            marked_text: indoc! {r#"
-                [fn other() { }
-
-                fn long_function() {
-                    let line1 = 1;
-                    let line2 = 2;
-                    let line3 = 3;
-                    let line4 = 4;
-                    let line5 = 5;
-                    let line6 = 6;
-                «    let line7 = 7;
-                    let line8 = 8;
-                    let line9 = 9;
-                    let line10 = 10;ˇ
-                    let line11 = 11;
-                }
-
-                fn another() { }»]
-            "#},
-            // 40 tokens for editable - allows several lines plus syntax expansion
-            editable_token_limit: 40,
-            // 55 tokens - enough for function but not whole file
-            context_token_limit: 55,
-        },
-    ];
-
-    for test_case in test_cases {
-        let cursor_marker: TextRangeMarker = 'ˇ'.into();
-        let editable_marker: TextRangeMarker = ('«', '»').into();
-        let context_marker: TextRangeMarker = ('[', ']').into();
-
-        let (text, mut ranges) = marked_text_ranges_by(
-            test_case.marked_text,
-            vec![
-                cursor_marker.clone(),
-                editable_marker.clone(),
-                context_marker.clone(),
-            ],
-        );
-
-        let cursor_ranges = ranges.remove(&cursor_marker).unwrap_or_default();
-        let expected_editable = ranges.remove(&editable_marker).unwrap_or_default();
-        let expected_context = ranges.remove(&context_marker).unwrap_or_default();
-        assert_eq!(expected_editable.len(), 1);
-        assert_eq!(expected_context.len(), 1);
-
-        cx.new(|cx| {
-            let text = text.trim_end_matches('\n');
-            let buffer = Buffer::local(text, cx).with_language(rust_lang(), cx);
-            let snapshot = buffer.snapshot();
-
-            let cursor_offset = cursor_ranges[0].start;
-            let cursor_point = snapshot.offset_to_point(cursor_offset);
-            let expected_editable_start = snapshot.offset_to_point(expected_editable[0].start);
-            let expected_editable_end = snapshot.offset_to_point(expected_editable[0].end);
-            let expected_context_start = snapshot.offset_to_point(expected_context[0].start);
-            let expected_context_end = snapshot.offset_to_point(expected_context[0].end);
-
-            let (actual_editable, actual_context) = editable_and_context_ranges_for_cursor_position(
-                cursor_point,
-                &snapshot,
-                test_case.editable_token_limit,
-                test_case.context_token_limit,
-            );
-
-            let range_text = |start: Point, end: Point| -> String {
-                snapshot.text_for_range(start..end).collect()
-            };
-
-            let editable_match = actual_editable.start == expected_editable_start
-                && actual_editable.end == expected_editable_end;
-            let context_match = actual_context.start == expected_context_start
-                && actual_context.end == expected_context_end;
-
-            if !editable_match || !context_match {
-                println!("\n=== FAILED: {} ===", test_case.name);
-                if !editable_match {
-                    println!(
-                        "\nExpected editable ({:?}..{:?}):",
-                        expected_editable_start, expected_editable_end
-                    );
-                    println!(
-                        "---\n{}---",
-                        range_text(expected_editable_start, expected_editable_end)
-                    );
-                    println!(
-                        "\nActual editable ({:?}..{:?}):",
-                        actual_editable.start, actual_editable.end
-                    );
-                    println!(
-                        "---\n{}---",
-                        range_text(actual_editable.start, actual_editable.end)
-                    );
-                }
-                if !context_match {
-                    println!(
-                        "\nExpected context ({:?}..{:?}):",
-                        expected_context_start, expected_context_end
-                    );
-                    println!(
-                        "---\n{}---",
-                        range_text(expected_context_start, expected_context_end)
-                    );
-                    println!(
-                        "\nActual context ({:?}..{:?}):",
-                        actual_context.start, actual_context.end
-                    );
-                    println!(
-                        "---\n{}---",
-                        range_text(actual_context.start, actual_context.end)
-                    );
-                }
-                panic!("Test '{}' failed - see output above", test_case.name);
-            }
-
-            buffer
-        });
-    }
 }
